Slight progress
Created an collision object and a collision script that should be able
to display text once added to an object. Unfortunately I cannot seem to
add the script to the object

In other news, I tried changing the rigidbody movement to get rid of the
bug, but was unable to do so. I did as you said and created the 'if key
released' functions, however that seemed to lock up the movement
functions even sooner (sometimes locking up after I press only one key).
I looked at some of Xenko's sample code for the first person shooter and
'New Game' options, but they used very similar code to yours. They do
have some extensions that we can use for movement and key mapping. We
currently do not have those extensions.